Move part of config into new dataset package
