+++ pkg/datasets/config.go
@@ -1,4 +1,4 @@
-package config
+package datasets
 
 import (
 	"embed"
